HttpServerCommand now provides a default Startup class implementation
