+++ src/AppMotor.HttpServer/AppMotor.HttpServer.csproj
@@ -23,6 +23,10 @@
   </ItemGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.Razor.RuntimeCompilation" Version="5.0.6" />
+  </ItemGroup>
+
+  <ItemGroup>
     <ProjectReference Include="..\AppMotor.CliApp.Hosting\AppMotor.CliApp.Hosting.csproj" />
     <ProjectReference Include="..\AppMotor.Core\AppMotor.Core.csproj" />
   </ItemGroup>
